pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sits"
version = "0.1.0"
description = "Satellite Image Time Series - Clustering and Classification"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SITS Team"}
]
keywords = ["satellite", "time-series", "clustering", "classification", "remote-sensing"]

dependencies = [
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "rasterio>=1.3.0",
    "scikit-learn>=1.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "tqdm>=4.65.0",
    "tsai>=0.3.9",
]

[project.optional-dependencies]
annotation = [
    "PyQt6>=6.5.0",
    "pyqtgraph>=0.13.0",
    "PyYAML>=6.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]
all = [
    "sits[annotation]",
    "sits[dev]",
]

[project.scripts]
sits-annotate = "sits.annotation.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sits"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
